requirements.txt
# MCP Server Dependencies
requests==2.31.0
flask==3.0.0
python-dotenv==1.0.0

# Backend Agent Dependencies
flask-cors==4.0.0
langchain-core==0.2.43
langchain-anthropic==0.1.15
anthropic==0.29.0
pydantic>=2.7.4
